+++ sql-to-dbsp-compiler/SQL-compiler/src/test/java/org/dbsp/sqlCompiler/compiler/sql/MetadataTests.java
@@ -1549,7 +1549,8 @@ public class MetadataTests extends BaseSQLTests {
                       "unused" : false
                     } ],
                     "materialized" : false
-                  } ]
+                  } ],
+                  "sql_settings" : { }
                 }""", jsonContents);
     }